fix unit tests after renaming
