+++ tests/FamilyHubs.ReferralUi.UnitTests/Pages/ProfessionalReferral/WhenUsingSupportDetails.cs
@@ -42,7 +42,7 @@ public class WhenUsingSupportDetails : BaseProfessionalReferralPage
 
         //Assert
         ArgumentNullException.ThrowIfNull(result);
-        result.PageName.Should().Be("/ConnectionRequest/WhySupport");
+        result.PageName.Should().Be("/ProfessionalReferral/WhySupport");
     }
 
     [Theory]
