+++ tests/FamilyHubs.ReferralUi.UnitTests/Pages/ProfessionalReferral/WhenUsingWhySupport.cs
@@ -5,7 +5,7 @@ using FluentAssertions;
 using Microsoft.AspNetCore.Mvc;
 using Moq;
 
-namespace FamilyHubs.ReferralUi.UnitTests.Pages.ConnectionRequest;
+namespace FamilyHubs.ReferralUi.UnitTests.Pages.ProfessionalReferral;
 
 public class WhenUsingWhySupport
 {
